prepare for upload to server

+++ app/Http/Controllers/Admin/UserController.php
@@ -4,27 +4,31 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserRequest;
-use App\Models\Role;
 use App\Models\User;
 use App\Services\Actions\User\GetList;
+use App\Services\Actions\User\RoleAndPermissionLevelAccess;
+use App\Services\Cache\CategoryCache;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
     public function index()
     {
-        $this->authorizeRoleOrPermission('view users');
+        $this->authorizeRoleOrPermission('show users');
         $users = GetList::handle();
+
         return view('admin.users.index', compact('users'));
     }
 
     public function create()
     {
         $this->authorizeRoleOrPermission('create users');
-        $roles = Role::all();
-        return view('admin.users.create', compact('roles'));
+
+        $roles = (new RoleAndPermissionLevelAccess())->getRolesByAccessLevel(auth()->user());
+        $categories = CategoryCache::allActive(config('pars-ticket.cache.timeout-long'));
+
+        return view('admin.users.create', compact('roles', 'categories'));
     }
 
     public function store(UserRequest $request)
@@ -35,23 +39,37 @@ class UserController extends Controller
         $validated['password'] = Hash::make($validated['password']);
 
         $user = User::create($validated);
-        $user->assignRole($request->roles);
+        if ($request->has('categories')) {
+            $user->categories()->sync($request->categories);
+        }
 
+        if ($this->canAuthorizeRoleOrPermission('update users roles')) {
+            $user->assignRole($request->roles);
+        }
 
         return redirect()->route('admin.users.index')
-            ->with('success', 'کاربر با موفقیت ایجاد شد.');
+            ->with('success', __('user.created_user_success'));
     }
 
     public function edit(User $user)
     {
-        $this->authorizeRoleOrPermission('edit users');
-        $roles = Role::all();
-        return view('admin.users.create', compact('user', 'roles'));
+        $this->authorizeRoleOrPermission('update users');
+
+        $user->load('categories');
+        $roles = (new RoleAndPermissionLevelAccess())->getRolesByAccessLevel(auth()->user());
+        $categories = CategoryCache::allActive(config('pars-ticket.cache.timeout-long'));
+
+        return view('admin.users.create', compact('user', 'roles', 'categories'));
     }
 
     public function update(UserRequest $request, User $user)
     {
-        $this->authorizeRoleOrPermission('edit users');
+        $this->authorizeRoleOrPermission('update users');
+
+        if (! (new RoleAndPermissionLevelAccess())->CheckRoleInUpdate(auth()->user(), $user)) {
+            return redirect()->route('admin.users.index')
+                ->with('error', __('user.you_can_not_edit_this_user_message'));
+        }
 
         $validated = $request->validationData();
 
@@ -62,18 +80,30 @@ class UserController extends Controller
         }
 
         $user->update($validated);
-        $user->syncRoles($request->roles);
+        if ($request->has('categories')) {
+            $user->categories()->sync($request->categories);
+        }
+
+        if ($this->canAuthorizeRoleOrPermission('update users roles')) {
+            $user->syncRoles($request->roles);
+        }
 
         return redirect()->route('admin.users.index')
-            ->with('success', 'کاربر با موفقیت بروزرسانی شد.');
+            ->with('success', __('user.user_updated_success_message'));
     }
 
     public function destroy(User $user)
     {
         $this->authorizeRoleOrPermission('delete users');
+
+        if (! (new RoleAndPermissionLevelAccess())->CheckRoleInUpdate(auth()->user(), $user)) {
+            return redirect()->route('admin.users.index')
+                ->with('error', __('user.you_can_not_deleted_user_message'));
+        }
+
         $user->delete();
 
         return redirect()->route('admin.users.index')
-            ->with('success', 'کاربر با موفقیت حذف شد.');
+            ->with('success', __('user.user_deleted_success_message'));
     }
 }
